+++ src/importgraph.c
@@ -6,14 +6,9 @@
 
 #include "importgraph.h"
 #include "defs.h"
+#include "file.h"
 #include "lib/vector.h"
 #include "lib/map.h"
-
-static void FileCleanup(FILE **);
-static cstring *LoadFile(const cstring *);
-static cstring *cStringFromFile(FILE *);
-static size_t GetFileSize(FILE *);
-static xerror Insert(map(Vertex) *, const cstring *);
 
 //------------------------------------------------------------------------------
 // the finite import graph is implemented as an adjacency list. Each vertex is
@@ -93,7 +88,7 @@ static file *InsertVertex(map(Vertex) *import_graph, const cstring *fname)
 		return root;
 	}
 
-	cstring *source_code = LoadFile(fname);
+	cstring *source_code = FileLoad(fname);
 
 	if (!source_code) {
 		exception = XXIO;
@@ -149,91 +144,4 @@ static InsertNeighbors(map(Vertex) *import_graph, vector(Import) imports)
 		import *node = &imports.buffer[i];
 		node->root = InsertVertex(import_graph, node.alias);
 	}
-}
-
-//------------------------------------------------------------------------------
-// file management
-
-static void FileCleanup(FILE **handle)
-{
-	if (*handle) {
-		fclose(*handle);
-	}
-}
-
-//on failure returns NULL, else returns a dynamically allocated cstring. Any
-//errors are reported to the xerror log.
-static cstring *LoadFile(const cstring *filename)
-{
-	assert(filename);
-
-	RAII(FileCleanup) FILE *fp = fopen(fname, "r");
-
-	if (!fp) {
-		xerror_issue("%s: %s", fname, strerror(errno));
-		return NULL;
-	}
-
-	cstring *src = cStringFromFile(fp);
-
-	if (!src) {
-		const cstring *msg = XerrorDescription(*err);
-		xerror_issue("%s: %s: cannot copy file to memory", fname, msg);
-	}
-
-	return src;
-}
-
-//on failure returns NULL, else returns a dynamically allocated cstring
-static cstring *cStringFromFile(FILE *openfile)
-{
-	assert(openfile);
-	assert(err);
-
-	size_t filesize = GetFileSize(openfile);
-
-	if (!filesize) {
-		xerror_issue("cannot calculate file size");
-		return NULL;
-	}
-
-	const size_t buflen = sizeof(char) * filesize + 1;
-	cstring *buffer = AbortMalloc(buflen);
-
-	size_t total_read = fread(buffer, sizeof(char), filesize, openfile);
-
-	if (total_read != filesize) {
-		xerror_issue("%s", strerror(errno));
-		return NULL;
-	}
-
-	buffer[filesize] = '\0';
-	return buffer;
-}
-
-//on failure returns zero, the openfile position will be set to zero on success
-static size_t GetFileSize(FILE *openfile)
-{
-	assert(openfile);
-	assert(bytes);
-
-	rewind(openfile);
-
-	long err = fseek(openfile, 0L, SEEK_END);
-
-	if (err == -1) {
-		xerror_issue("fseek: cannot move to EOF: %s", strerror(errno));
-		return 0;
-	}
-
-	long bytecount = ftell(openfile);
-
-	if (bytecount == -1) {
-		xerror_issue("ftell: cannot fetch count: %s", strerror(errno));
-		return 0;
-	}
-
-	rewind(openfile);
-
-	return (size_t) bytecount;
 }
